Revert "Added classes and tests to app"
