+++ Gemfile
@@ -2,4 +2,3 @@ source 'https://rubygems.org'
 gem 'nokogiri'
 gem 'rack', '~> 2.0.1'
 gem 'rspec'
-gem 'sinatra'
