+++ index.html
@@ -1,4 +1,3 @@
-
 <!DOCTYPE html> 
 <html>
 <head> 
@@ -6,49 +5,19 @@
     <link rel="stylesheet" type="text/css" href="style.css"/>
     <link href="https://fonts.googleapis.com/css?family=Raleway:300,400" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Open+Sans|Rock+Salt" rel="stylesheet">
-    
-    
-<style>   
-    body {
-    background-color:indianred;
-}
-
-#add {
-    text-align: center;
-    font-family: "Raleway", sans-serif;
-    background-color: rgba(95, 10, 224, 0.78);
-    padding-bottom:100px;
-}
-
-textarea {
-    margin-left:10px;
-}
-
-#title {
-    color:white;
-    padding-top:15px;
-}
-
-#submit {
-    margin-bottom:10px;
-}
-
-#checklist {
-    background-color: white;
-}
-</style>    
-    
 </head>    
 <body>
-    <form id="add" method="post" action="/submit/">
+    <form id="add">
         <h1 id="title">Your To Do List</h1>
         <textarea id="subject" name="listItem" placeholder="Add an item to the list.." style="height:50px; width:300px;"></textarea>
         <br/>
-        <input value="submit" type="submit"/>
+        <button id="submit">Submit</button>
    <br/>
+    <form id="checklist">
+        <input type="checkbox" name="vehicle" value="Bike"> I have a bike<br>
     </form>
     
-    <div id="list" name="list" style="width:200px; height:200px;"></div>
+    </form>
     
 </body>
 </html>
