+++ style.css
@@ -1,11 +1,11 @@
 body {
     background-color:indianred;
-    margin: 0 auto;
 }
 
-form {
+#add {
     text-align: center;
     font-family: "Raleway", sans-serif;
+    background-color: rgba(95, 10, 224, 0.78);
     padding-bottom:100px;
 }
 
@@ -14,8 +14,6 @@ textarea {
 }
 
 #title {
-    text-align: center;
-    font-family: "Raleway", sans-serif;
     color:white;
     padding-top:15px;
 }
@@ -24,9 +22,6 @@ textarea {
     margin-bottom:10px;
 }
 
-.newList{
-    text-align: center;
-    font-family: "Raleway", sans-serif;
-    margin: 0 auto;
+#checklist {
+    background-color: white;
 }
-
