+++ to_do_list.rb
@@ -10,18 +10,15 @@ class Item
     def complete_task
         @completed = true
     end
-    
     def description
-        @description
     end
 end
 
 
 class ToDoList
     
-    def initialize(display)
+    def initialize()
        @to_do_list = []
-       @display = display
     end
     
     def add_task(item)
@@ -29,8 +26,7 @@ class ToDoList
     end
     
     def print_list()
-        list_of_items = @to_do_list.map {|item| item.description}
-        @display.present(list_of_items)
+        @to_do_list.each {|item| puts item.description}
     end
     
     def remove_task(item)
@@ -39,54 +35,36 @@ class ToDoList
         
 end
 
-
 class Engine 
-    def initialize(to_do_list, display)
+    def initialize(to_do_list)
       @to_do_list = to_do_list
-      @display = display
     end
     
-    def user_response
+    def add_one
         user_prompts
-        response = @display.receive
-        while response == "yes"
-        user_prompts
-        response = @display.receive    
-        end 
-        @to_do_list.print_list()  
+            response = gets.chomp
+            while response == "yes"
+            user_prompts
+            response = gets.chomp    
+            end 
+            @to_do_list.print_list()  
     end
     
     def user_prompts
-        @display.present("Add an item to your to do list")
-        thing = @display.receive
-        item = Item.new(thing)
+        puts "Add an item to your to do list"
+        item = Item.new(gets.chomp)
         @to_do_list.add_task(item) 
         @to_do_list.print_list()
-        @display.present("Do you want to add another item? Yes or no?")
-    end    
-    
-end
-
-class Display
-  def initialize(output = $stdout, input = $stdin)
-    @output = output
-    @input = input
-  end
-    
-  def present(message)
-    @output.puts message
-  end
-    
-  def receive 
-    @input.gets.chomp
-  end
+        puts "Do you want to add another item? Yes or no?"
+    end
 end
         
-    display = Display.new  
-    to_do_list = ToDoList.new(display)
-    engine = Engine.new(to_do_list, display)
-    engine.user_response
+        
+    to_do_list = ToDoList.new
+    engine = Engine.new(to_do_list)
+    engine.plus_one
     
+
