--- views/lists/new.erb
@@ -1,15 +1,0 @@
-<html>
-    <head>
-        <link rel="stylesheet" href="/style.css"/>
-        <link href="https://fonts.googleapis.com/css?family=Raleway:300,400" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/css?family=Open+Sans|Rock+Salt" rel="stylesheet">
-    </head>
-<body>
-    <h1 id="title">Your To Do List</h1>
-        <form method="post" action="/lists/new">
-            <input type="text" name="name" placeholder="List Name"><br><br>
-            <input type="text" name="tasks[name]" placeholder="Add Your First Task"><br><br>
-            <input class="btn btn-primary" type="submit" name="create" value="Create">
-        </form>
-</body>
-</html>
