--- views/tasks/home.erb
@@ -1,19 +1,0 @@
-<html>
-<head>
-<link rel="stylesheet" href="/style.css"/>
-<link href="https://fonts.googleapis.com/css?family=Raleway:300,400" rel="stylesheet">
-<link href="https://fonts.googleapis.com/css?family=Open+Sans|Rock+Salt" rel="stylesheet">
-</head>
-<body>
-    
-    <h1 id="title">Your To Do List</h1>
-    
-    <div class="newList">
-    <a href="/lists/new"><button type="button">Create A New List</button></a>
-    </div>
-    
-    
-    <div id="list" name="list" style="width:200px; height:200px;"></div>
-    
-</body>
-</html>
