--- views/tasks/layout.erb
@@ -1,12 +1,0 @@
-<!doctype html>
-<html>
-<head>
- <title>To Do List Application</title>
-        
-</head>
-<body>
-    <h1 id="title">Your To Do List</h1>
-    
-    
-</body>
-</html>
